added some more Unit tests.
